Setup barebones metadata schema
The `GeneratedResourceSchemaStrategy` implements an
`ActiveFedora::SchemaIndexingStrategy` that works around application order limitations in
`ActiveFedora::GeneratedResourceSchema`. This serves as a prophylactic against
one of the more common tripping points in Samvera metadata work.

Adding properties to `Schemas::WorkMetadata` instead of directly to the `Work`
model will ensure they are applied according to the custom strategy.