Alias `Work#rights` attribute to `#license`
The property `#rights` has been renamed `#license` since `CurationConcerns`
1.7.x. Since to-be-ported code may refer to this property as `#rights` this
alias will ease porting that code by offering model parity across versions.

We may want to change the display labels to `"Rights"` as well.

The document at
https://docs.google.com/spreadsheets/d/1StB1cVRobKHqQmu47bnt8adJjA3xwSN8aE5xZiQYHG8/
appears to be incorrect concerning the implementation of this property in the
legacy Sufia application. This should be reviewed as concerns #27.

This work advances #45.